refactor(frontend): decompose Dashboard.tsx into focused sub-components
Extract BudgetSpendingGauge → components/dashboard/BudgetSection.tsx
Extract RecentTransactionsFeed → components/dashboard/RecentTransactionsSection.tsx
Extract tagging rule evaluation → utils/taggingRuleEval.ts
Create canonical Transaction type → types/transaction.ts

Dashboard.tsx reduced from ~1900 to ~1000 lines. All functionality preserved.
