+++ frontend/src/components/TransactionsTable.tsx
@@ -38,22 +38,9 @@ import { useCategoryTagCreate } from "../hooks/useCategoryTagCreate";
 import { useCategories } from "../hooks/useCategories";
 import { useCashBalances } from "../hooks/useCashBalances";
 import { useTranslation } from "react-i18next";
+import type { Transaction } from "../types/transaction";
 
-export interface Transaction {
-  id?: number;
-  unique_id?: string;
-  source?: string;
-  desc?: string;
-  description?: string;
-  amount: number;
-  date: string;
-  category?: string;
-  tag?: string;
-  provider?: string;
-  account_name?: string;
-  account_number?: string;
-  pending_refund_id?: number; // ID if this transaction has a pending refund
-}
+export type { Transaction };
 
 export interface TransactionsTableProps {
   transactions: Transaction[];
