+++ frontend/src/pages/Dashboard.tsx
@@ -1,17 +1,9 @@
-import { useState, useMemo, useRef, useCallback, useEffect } from "react";
-import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
-import { Link } from "react-router-dom";
+import { useState, useMemo } from "react";
+import { useQuery } from "@tanstack/react-query";
 import {
   TrendingDown,
-  ChevronLeft,
-  ChevronRight,
   Calculator,
-  Split,
-  RefreshCw,
   Tag,
-  Wand2,
-  Link2,
-  Plus,
 } from "lucide-react";
 import Plot from "react-plotly.js";
 import {
@@ -19,31 +11,18 @@ import {
   cashBalancesApi,
   bankBalancesApi,
   investmentsApi,
-  budgetApi,
   transactionsApi,
   taggingApi,
-  pendingRefundsApi,
-  type TaggingRule,
-  type ConditionNode,
   type BankBalance,
 } from "../services/api";
-import { SplitTransactionModal } from "../components/modals/SplitTransactionModal";
-import { LinkRefundModal } from "../components/modals/LinkRefundModal";
-import { ProjectModal } from "../components/modals/ProjectModal";
-import { SelectDropdown } from "../components/common/SelectDropdown";
-import { useCategoryTagCreate } from "../hooks/useCategoryTagCreate";
-import { useCategories } from "../hooks/useCategories";
-import { useTaggingRules } from "../hooks/useTaggingRules";
+import { BudgetSpendingGauge } from "../components/dashboard/BudgetSection";
+import { RecentTransactionsFeed } from "../components/dashboard/RecentTransactionsSection";
 import { SankeyChart } from "../components/SankeyChart";
-import { SemiGauge } from "../components/common/SemiGauge";
 import { Skeleton } from "../components/common/Skeleton";
 import { useDemoMode } from "../context/DemoModeContext";
-import { isToday, isYesterday } from "date-fns";
 import { useTranslation } from "react-i18next";
-import { formatMonthYear, formatShortDate } from "../utils/dateFormatting";
 import { formatCurrency, formatCompactCurrency } from "../utils/numberFormatting";
 import { chartTheme, plotlyConfig, isTouchDevice } from "../utils/plotlyLocale";
-import i18n from "../i18n";
 
 type NetWorthView = "all" | "bank_balance" | "investments" | "net_worth" | "debt_payments";
 
@@ -186,896 +165,12 @@ function FinancialHealthHeader({
   );
 }
 
-/* ------------------------------------------------------------------ */
-/*  Section 2 — Budget Spending Gauge (Monthly + Projects)            */
-/* ------------------------------------------------------------------ */
-
-type BudgetViewMode = "monthly" | "projects";
-
-interface BudgetRule {
-  id: number;
-  name: string;
-  category: string;
-  tags: string | null;
-  budget_amount: number;
-  spent_amount: number;
-}
-
-function getProgressColor(pct: number): string {
-  if (pct > 100) return "bg-rose-500";
-  if (pct >= 75) return "bg-amber-500";
-  return "bg-emerald-500";
-}
-
-function BudgetRuleCards({
-  rules,
-  categoryIcons,
-}: {
-  rules: BudgetRule[];
-  categoryIcons: Record<string, string> | undefined;
-}) {
-  const { t } = useTranslation();
-  return (
-    <div className="h-[260px] overflow-y-auto scrollbar-auto-hide mb-4">
-      {rules.length > 0 && (
-        <div className="grid grid-cols-2 gap-3">
-          {rules.map((rule) => {
-            const pct = rule.budget_amount > 0 ? (rule.spent_amount / rule.budget_amount) * 100 : 0;
-            const remaining = rule.budget_amount - rule.spent_amount;
-            const icon = categoryIcons?.[rule.category] ?? "";
-            return (
-              <div
-                key={rule.id}
-                className="bg-[var(--surface-light)] rounded-xl p-3.5 space-y-2"
-              >
-                <div className="flex items-center justify-between">
-                  <div className="flex items-center gap-1.5 min-w-0">
-                    {icon && <span className="text-sm flex-shrink-0">{icon}</span>}
-                    <span className="text-xs font-semibold truncate">{rule.name}</span>
-                  </div>
-                  <span className={`text-[10px] font-bold px-1.5 py-0.5 rounded-full flex-shrink-0 ${
-                    pct > 100
-                      ? "bg-rose-500/20 text-rose-400"
-                      : pct >= 75
-                        ? "bg-amber-500/20 text-amber-400"
-                        : "bg-emerald-500/20 text-emerald-400"
-                  }`}>
-                    {Math.round(pct)}%
-                  </span>
-                </div>
-                <p className="text-sm font-bold">
-                  {formatCurrency(rule.spent_amount)}
-                  <span className="text-xs font-normal text-[var(--text-muted)]">
-                    {" "}/ {formatCurrency(rule.budget_amount)}
-                  </span>
-                </p>
-                <div className="h-1.5 w-full rounded-full bg-[var(--surface)] overflow-hidden">
-                  <div
-                    className={`h-full rounded-full transition-all ${getProgressColor(pct)}`}
-                    style={{ width: `${Math.min(pct, 100)}%` }}
-                  />
-                </div>
-                <p className={`text-[10px] font-medium ${
-                  remaining >= 0 ? "text-[var(--text-muted)]" : "text-rose-400"
-                }`}>
-                  {remaining >= 0
-                    ? `${formatCurrency(remaining)} ${t("budget.remaining").toLowerCase()}`
-                    : `${formatCurrency(Math.abs(remaining))} ${t("budget.overBudget").toLowerCase()}`}
-                </p>
-              </div>
-            );
-          })}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function BudgetSpendingGauge({
-  categoryIcons,
-}: {
-  categoryIcons: Record<string, string> | undefined;
-}) {
-  const { t } = useTranslation();
-  const now = new Date();
-  const currentYear = now.getFullYear();
-  const currentMonth = now.getMonth() + 1;
-  const [year, setYear] = useState(currentYear);
-  const [month, setMonth] = useState(currentMonth);
-  const [viewMode, setViewMode] = useState<BudgetViewMode>("monthly");
-  const [selectedProject, setSelectedProject] = useState<string | null>(null);
-  const [isProjectModalOpen, setIsProjectModalOpen] = useState(false);
-  const { isDemoMode } = useDemoMode();
-
-  const isCurrentMonth = year === currentYear && month === currentMonth;
-  const monthDate = new Date(year, month - 1);
-  const monthName = formatMonthYear(monthDate);
-  const daysInMonth = new Date(year, month, 0).getDate();
-  const daysRemaining = daysInMonth - now.getDate();
-
-  const handlePreviousMonth = () => {
-    if (month === 1) {
-      setMonth(12);
-      setYear(year - 1);
-    } else {
-      setMonth(month - 1);
-    }
-  };
-
-  const handleNextMonth = () => {
-    if (month === 12) {
-      setMonth(1);
-      setYear(year + 1);
-    } else {
-      setMonth(month + 1);
-    }
-  };
-
-  const queryClient = useQueryClient();
-
-  // Prefetch surrounding 11 months so navigation is instant
-  useEffect(() => {
-    for (let i = 1; i <= 11; i++) {
-      const d = new Date(currentYear, currentMonth - 1 - i);
-      const prefetchYear = d.getFullYear();
-      const prefetchMonth = d.getMonth() + 1;
-      queryClient.prefetchQuery({
-        queryKey: ["budget-analysis", prefetchYear, prefetchMonth, isDemoMode],
-        queryFn: async () => {
-          const res = await budgetApi.getAnalysis(prefetchYear, prefetchMonth, false);
-          return res.data;
-        },
-      });
-    }
-  }, [isDemoMode, currentYear, currentMonth, queryClient]);
-
-  // --- Monthly budget data ---
-  const { data: rawBudgetAnalysis, isLoading: monthlyLoading } = useQuery({
-    queryKey: ["budget-analysis", year, month, isDemoMode],
-    queryFn: async () => {
-      const res = await budgetApi.getAnalysis(year, month, false);
-      return res.data;
-    },
-    enabled: viewMode === "monthly",
-  });
-
-  const budgetAnalysis = useMemo(() => {
-    if (!rawBudgetAnalysis?.rules) return undefined;
-    const rules: BudgetRule[] = rawBudgetAnalysis.rules.map((item: { rule: { id: number; name: string; category: string; tags: string | null; amount: number }; current_amount: number }) => ({
-      id: item.rule.id,
-      name: item.rule.name,
-      category: item.rule.category,
-      tags: item.rule.tags,
-      budget_amount: item.rule.amount,
-      spent_amount: item.current_amount,
-    }));
-    const totalRule = rules.find((r) => r.name === "Total Budget");
-    return {
-      rules,
-      total_budget: totalRule?.budget_amount,
-      total_spent: totalRule?.spent_amount,
-    };
-  }, [rawBudgetAnalysis]);
-
-  // --- Project budget data ---
-  const { data: projects } = useQuery({
-    queryKey: ["budget-projects", isDemoMode],
-    queryFn: async () => {
-      const res = await budgetApi.getProjects();
-      return res.data as string[];
-    },
-    enabled: viewMode === "projects",
-  });
-
-  const createProjectMutation = useMutation({
-    mutationFn: budgetApi.createProject,
-    onSuccess: (_, variables) => {
-      queryClient.invalidateQueries({ queryKey: ["budget-projects"] });
-      queryClient.invalidateQueries({ queryKey: ["availableProjects"] });
-      setSelectedProject(variables.category);
-      setIsProjectModalOpen(false);
-    },
-  });
-
-  // Auto-select first project when projects load
-   
-  useEffect(() => {
-    if (projects && projects.length > 0 && !selectedProject) {
-      // eslint-disable-next-line react-hooks/set-state-in-effect
-      setSelectedProject(projects[0]);
-    }
-  }, [projects, selectedProject]);
-
-  const { data: rawProjectDetails, isLoading: projectLoading } = useQuery({
-    queryKey: ["budget-project-details", selectedProject, isDemoMode],
-    queryFn: async () => {
-      const res = await budgetApi.getProjectDetails(selectedProject!, false);
-      return res.data;
-    },
-    enabled: viewMode === "projects" && !!selectedProject,
-  });
-
-  const projectAnalysis = useMemo(() => {
-    if (!rawProjectDetails?.rules) return undefined;
-    const rules: BudgetRule[] = rawProjectDetails.rules.map((item: { rule: { id: number; name: string; category: string; tags: string | null; amount: number }; current_amount: number }) => ({
-      id: item.rule.id,
-      name: item.rule.name,
-      category: item.rule.category,
-      tags: item.rule.tags,
-      budget_amount: item.rule.amount,
-      spent_amount: item.current_amount,
-    }));
-    const totalRule = rules.find((r) => r.name === "Total Budget");
-    return {
-      rules,
-      total_budget: totalRule?.budget_amount,
-      total_spent: totalRule?.spent_amount ?? rawProjectDetails.total_spent,
-    };
-  }, [rawProjectDetails]);
-
-  // --- Computed values ---
-  const activeAnalysis = viewMode === "monthly" ? budgetAnalysis : projectAnalysis;
-  const isLoading = viewMode === "monthly" ? monthlyLoading : projectLoading;
-
-  const totalBudget =
-    activeAnalysis?.total_budget ??
-    (activeAnalysis?.rules?.reduce((sum, r) => sum + r.budget_amount, 0) ?? 0);
-  const totalSpent =
-    activeAnalysis?.total_spent ??
-    (activeAnalysis?.rules?.reduce((sum, r) => sum + r.spent_amount, 0) ?? 0);
-
-  const miniRules = useMemo(() => {
-    if (!activeAnalysis?.rules) return [];
-    return activeAnalysis.rules
-      .filter((r) => r.name !== "Total Budget")
-      .sort((a, b) => b.spent_amount - a.spent_amount);
-  }, [activeAnalysis]);
-
-  // For monthly: hide card entirely if no data at all
-  if (viewMode === "monthly" && !monthlyLoading && (!budgetAnalysis || budgetAnalysis.rules.length === 0)) return null;
-
-  const hasNoProjects = viewMode === "projects" && (!projects || projects.length === 0);
-
-  const toggleViewMode = () =>
-    setViewMode(viewMode === "monthly" ? "projects" : "monthly");
-
-  const segmentedControlEl = (
-    <button
-      onClick={toggleViewMode}
-      className="flex bg-[var(--surface-light)] p-0.5 rounded-lg cursor-pointer"
-    >
-      {([
-        { key: "monthly" as const, label: t("budget.monthlyBudget") },
-        { key: "projects" as const, label: t("budget.projectBudgets") },
-      ]).map(({ key, label }) => (
-        <span
-          key={key}
-          className={`px-3 py-1 rounded-md text-xs font-semibold transition-all ${
-            viewMode === key
-              ? "bg-[var(--surface)] text-[var(--primary)] shadow-sm"
-              : "text-[var(--text-muted)]"
-          }`}
-        >
-          {label}
-        </span>
-      ))}
-    </button>
-  );
-
-  return (
-    <div
-      className="bg-[var(--surface)] rounded-2xl p-4 md:p-6 border border-[var(--surface-light)]"
-    >
-      {/* Header row: segmented control */}
-      <div className="flex items-center justify-between mb-4">
-        <p className="text-xs font-semibold uppercase tracking-wider text-[var(--text-muted)]">
-          🎯 {t("budget.title")}
-        </p>
-        {segmentedControlEl}
-      </div>
-
-      {isLoading ? (
-        <>
-          <div className="h-9 mb-4">
-            <Skeleton variant="text" lines={1} className="h-full" />
-          </div>
-          <Skeleton variant="chart" className="h-40" />
-        </>
-      ) : (
-        <>
-          {/* Sub-header: month nav or project selector — fixed h-9 so gauge stays in place */}
-          <div className="h-9 flex items-center mb-4">
-            {viewMode === "monthly" ? (
-              <div className="flex items-center justify-between w-full">
-                <div className="flex items-center gap-2">
-                  <button
-                    onClick={handlePreviousMonth}
-                    className="p-1 rounded-lg hover:bg-[var(--surface-light)] text-[var(--text-muted)] hover:text-[var(--text)] transition-colors"
-                  >
-                    <ChevronLeft size={16} />
-                  </button>
-                  <p className="text-xs font-semibold uppercase tracking-wider text-[var(--text-muted)] w-36 text-center">
-                    {monthName}
-                  </p>
-                  <button
-                    onClick={handleNextMonth}
-                    disabled={isCurrentMonth}
-                    className="p-1 rounded-lg hover:bg-[var(--surface-light)] text-[var(--text-muted)] hover:text-[var(--text)] transition-colors disabled:opacity-30 disabled:pointer-events-none"
-                  >
-                    <ChevronRight size={16} />
-                  </button>
-                </div>
-                {isCurrentMonth && (
-                  <span className="text-xs text-[var(--text-muted)]">
-                    ⏳ {t("dashboard.daysRemaining", { count: daysRemaining })}
-                  </span>
-                )}
-              </div>
-            ) : (
-              !hasNoProjects && (
-                <div className="w-full flex items-center gap-2">
-                  <div className="flex-1">
-                    <SelectDropdown
-                      options={(projects ?? []).map((p) => ({ label: p, value: p }))}
-                      value={selectedProject ?? ""}
-                      onChange={setSelectedProject}
-                      placeholder="Select project..."
-                      size="sm"
-                    />
-                  </div>
-                  <button
-                    onClick={() => setIsProjectModalOpen(true)}
-                    className="p-1.5 rounded-lg hover:bg-[var(--surface-light)] text-[var(--primary)] transition-colors shrink-0"
-                    title={t("tooltips.addNewProject")}
-                  >
-                    <Plus size={16} />
-                  </button>
-                </div>
-              )
-            )}
-          </div>
-
-          {/* Empty state for projects */}
-          {hasNoProjects ? (
-            <div className="flex flex-col items-center justify-center py-12 text-center">
-              <p className="text-sm text-[var(--text-muted)] mb-3">{t("dashboard.noProjectBudgets")}</p>
-              <button
-                onClick={() => setIsProjectModalOpen(true)}
-                className="flex items-center gap-2 text-sm font-medium text-[var(--primary)] hover:text-[var(--primary-dark)] transition-colors cursor-pointer"
-              >
-                <Plus size={16} />
-                {t("budget.addProject")}
-              </button>
-            </div>
-          ) : (
-            <>
-              {/* Gauge */}
-              <div className="flex justify-center mb-6">
-                <SemiGauge spent={totalSpent} budget={totalBudget} size={240} />
-              </div>
-
-              {/* Budget rule cards */}
-              <BudgetRuleCards rules={miniRules} categoryIcons={categoryIcons} />
-
-              {/* Link to budget page */}
-              <div className="text-end">
-                <Link
-                  to="/budget"
-                  className="text-sm font-medium text-[var(--primary)] hover:underline"
-                >
-                  {t("dashboard.viewAllBudgetRules")} &rarr;
-                </Link>
-              </div>
-            </>
-          )}
-        </>
-      )}
-
-      <ProjectModal
-        isOpen={isProjectModalOpen}
-        onClose={() => setIsProjectModalOpen(false)}
-        onSubmit={(data) => createProjectMutation.mutate(data)}
-      />
-    </div>
-  );
-}
-
-/* ------------------------------------------------------------------ */
-/*  Section 3 — Recent Transactions Feed                              */
-/* ------------------------------------------------------------------ */
-
-interface Transaction {
-  id?: number;
-  unique_id?: string;
-  date: string;
-  description: string;
-  amount: number;
-  category: string;
-  tag: string;
-  source: string;
-  provider: string;
-  account_name: string;
-  pending_refund_id?: number;
-}
-
-function formatTransactionDate(dateStr: string): string {
-  const d = new Date(dateStr);
-  if (isToday(d)) return i18n.t("common.today");
-  if (isYesterday(d)) return i18n.t("common.yesterday");
-  return formatShortDate(d);
-}
-
-const TRANSACTIONS_PAGE_SIZE = 20;
-
-/** Evaluate a single condition leaf against a transaction. */
-function evalCondition(node: ConditionNode, tx: Transaction): boolean {
-  const { field, operator, value } = node;
-  if (!field || !operator) return false;
-
-  // Service field: compare against source
-  if (field === "service") {
-    if (operator !== "equals") return false;
-    const svc = String(value).toLowerCase().replace(/\s+/g, "_");
-    return (tx.source || "").toLowerCase().includes(svc);
-  }
-
-  const fieldMap: Record<string, string | number> = {
-    description: tx.description ?? "",
-    amount: tx.amount,
-    provider: tx.provider ?? "",
-    account_name: tx.account_name ?? "",
-  };
-  const fv = fieldMap[field];
-  if (fv === undefined) return false;
-
-  // Numeric operators
-  if (field === "amount") {
-    const n = Number(fv);
-    if (operator === "gt") return n > Number(value);
-    if (operator === "lt") return n < Number(value);
-    if (operator === "gte") return n >= Number(value);
-    if (operator === "lte") return n <= Number(value);
-    if (operator === "equals") return n === Number(value);
-    if (operator === "between" && Array.isArray(value))
-      return n >= Number(value[0]) && n <= Number(value[1]);
-    return false;
-  }
-
-  // Text operators (case-sensitive, matching backend LIKE behaviour)
-  const sv = String(fv);
-  const tv = String(value);
-  if (operator === "contains") return sv.includes(tv);
-  if (operator === "equals") return sv === tv;
-  if (operator === "starts_with") return sv.startsWith(tv);
-  if (operator === "ends_with") return sv.endsWith(tv);
-  return false;
-}
-
-/** Recursively evaluate a condition tree against a transaction. */
-function evalConditionTree(node: ConditionNode, tx: Transaction): boolean {
-  if (node.type === "CONDITION") return evalCondition(node, tx);
-  const subs = node.subconditions ?? [];
-  if (subs.length === 0) return true; // empty group matches all
-  if (node.type === "AND") return subs.every((s) => evalConditionTree(s, tx));
-  if (node.type === "OR") return subs.some((s) => evalConditionTree(s, tx));
-  return false;
-}
-
-/** Find the first tagging rule whose conditions match a transaction. */
-function findMatchingRule(
-  rules: TaggingRule[],
-  tx: Transaction,
-): TaggingRule | undefined {
-  return rules.find((r) => evalConditionTree(r.conditions, tx));
-}
-
-function RecentTransactionsFeed({
-  transactions,
-  categoryIcons,
-  isLoading,
-}: {
-  transactions: Transaction[] | undefined;
-  categoryIcons: Record<string, string> | undefined;
-  isLoading: boolean;
-}) {
-  const { t } = useTranslation();
-  const queryClient = useQueryClient();
-  const { createCategory, createTag } = useCategoryTagCreate();
-  const [visibleCount, setVisibleCount] = useState(TRANSACTIONS_PAGE_SIZE);
-  const sentinelRef = useRef<HTMLDivElement>(null);
-  const [editingTxKey, setEditingTxKey] = useState<string | null>(null);
-  const [mobileActionsTxKey, setMobileActionsTxKey] = useState<string | null>(null);
-  const [splittingTransaction, setSplittingTransaction] = useState<Transaction | null>(null);
-  const [linkingTransaction, setLinkingTransaction] = useState<Transaction | null>(null);
-
-  // Categories for inline tag editing
-  const { data: categories } = useCategories({ enabled: !!editingTxKey });
-  const { data: taggingRules } = useTaggingRules();
-
-  const invalidateAnalytics = useCallback(() => {
-    queryClient.invalidateQueries({ queryKey: ["income-outcome"] });
-    queryClient.invalidateQueries({ queryKey: ["analytics-category"] });
-    queryClient.invalidateQueries({ queryKey: ["sankey"] });
-    queryClient.invalidateQueries({ queryKey: ["net-worth-over-time"] });
-    queryClient.invalidateQueries({ queryKey: ["income-by-source"] });
-    queryClient.invalidateQueries({ queryKey: ["monthly-expenses"] });
-    queryClient.invalidateQueries({ queryKey: ["budget-analysis"] });
-  }, [queryClient]);
-
-  // Tag update mutation
-  const tagMutation = useMutation({
-    mutationFn: ({ tx, category, tag }: { tx: Transaction; category: string; tag: string }) =>
-      transactionsApi.updateTag(
-        String(tx.unique_id ?? tx.id),
-        category,
-        tag,
-        tx.source,
-      ),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["transactions"] });
-      queryClient.invalidateQueries({ queryKey: ["categories"] });
-      invalidateAnalytics();
-      setEditingTxKey(null);
-    },
-  });
-
-  // Mark as pending refund
-  const markPendingMutation = useMutation({
-    mutationFn: (tx: Transaction) =>
-      pendingRefundsApi.create({
-        source_type: "transaction",
-        source_id: tx.unique_id || "",
-        source_table: tx.source,
-        expected_amount: Math.abs(tx.amount),
-      }),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["transactions"] });
-      queryClient.invalidateQueries({ queryKey: ["pendingRefunds"] });
-      invalidateAnalytics();
-    },
-  });
-
-  const sorted = useMemo(() => {
-    if (!transactions) return [];
-    return [...transactions].sort(
-      (a, b) => new Date(b.date).getTime() - new Date(a.date).getTime(),
-    );
-  }, [transactions]);
-
-  const visible = useMemo(() => sorted.slice(0, visibleCount), [sorted, visibleCount]);
-  const hasMore = visibleCount < sorted.length;
-
-  // IntersectionObserver to auto-load more when sentinel enters viewport
-  const handleObserver = useCallback(
-    (entries: IntersectionObserverEntry[]) => {
-      const [entry] = entries;
-      if (entry.isIntersecting && hasMore) {
-        setVisibleCount((prev) => Math.min(prev + TRANSACTIONS_PAGE_SIZE, sorted.length));
-      }
-    },
-    [hasMore, sorted.length],
-  );
-
-  useEffect(() => {
-    const node = sentinelRef.current;
-    if (!node) return;
-    const observer = new IntersectionObserver(handleObserver, {
-      root: node.closest("[data-scroll-root]"),
-      rootMargin: "200px",
-    });
-    observer.observe(node);
-    return () => observer.disconnect();
-  }, [handleObserver]);
-
-  // Reset visible count when transactions change (e.g. demo mode toggle)
-   
-  useEffect(() => {
-    // eslint-disable-next-line react-hooks/set-state-in-effect
-    setVisibleCount(TRANSACTIONS_PAGE_SIZE);
-  }, [transactions]);
-
-  // Group by date label
-  const grouped = useMemo(() => {
-    const groups: { label: string; items: Transaction[] }[] = [];
-    let currentLabel = "";
-    for (const tx of visible) {
-      const label = formatTransactionDate(tx.date);
-      if (label !== currentLabel) {
-        groups.push({ label, items: [] });
-        currentLabel = label;
-      }
-      groups[groups.length - 1].items.push(tx);
-    }
-    return groups;
-  }, [visible]);
-
-  // Precompute which tagging rule matches each visible transaction
-  const ruleMatchMap = useMemo(() => {
-    if (!taggingRules || !visible.length) return new Map<string, TaggingRule>();
-    const map = new Map<string, TaggingRule>();
-    for (let i = 0; i < visible.length; i++) {
-      const tx = visible[i];
-      const key = `${tx.source}_${tx.unique_id ?? tx.id ?? `${tx.date}-${i}`}`;
-      const match = findMatchingRule(taggingRules, tx);
-      if (match) map.set(key, match);
-    }
-    return map;
-  }, [taggingRules, visible]);
-
-  const getTxKey = (tx: Transaction, i: number) =>
-    `${tx.source}_${tx.unique_id ?? tx.id ?? `${tx.date}-${i}`}`;
-
-  if (isLoading) {
-    return (
-      <div className="bg-[var(--surface)] rounded-2xl p-6 border border-[var(--surface-light)]">
-        <Skeleton variant="text" lines={1} className="mb-4" />
-        <Skeleton variant="text" lines={5} />
-      </div>
-    );
-  }
-
-  if (sorted.length === 0) return null;
-
-  return (
-    <div className="bg-[var(--surface)] rounded-2xl p-4 md:p-6 border border-[var(--surface-light)]">
-      <div className="flex items-center justify-between mb-4">
-        <p className="text-xs font-semibold uppercase tracking-wider text-[var(--text-muted)]">
-          🧾 {t("dashboard.recentTransactions")}
-        </p>
-        <Link
-          to="/transactions"
-          className="text-sm font-medium text-[var(--primary)] hover:underline"
-        >
-          {t("dashboard.viewAll")} &rarr;
-        </Link>
-      </div>
-
-      <div
-        data-scroll-root=""
-        className="max-h-[500px] overflow-y-auto space-y-4 scrollbar-auto-hide"
-      >
-        {grouped.map((group) => (
-          <div key={group.label}>
-            <p className="text-xs font-semibold text-[var(--text-muted)] mb-2 sticky top-0 bg-[var(--surface)] py-1 z-10">
-              {group.label}
-            </p>
-            <div className="space-y-1">
-              {group.items.map((tx, i) => {
-                const icon = categoryIcons?.[tx.category] ?? "";
-                const isPositive = tx.amount >= 0;
-                const txKey = getTxKey(tx, i);
-                const isEditing = editingTxKey === txKey;
-                const matchedRule = ruleMatchMap.get(txKey);
-
-                return (
-                  <div key={txKey}>
-                    <div
-                      className={`flex items-center gap-2 py-2 px-2 rounded-lg hover:bg-[var(--surface-light)]/40 transition-colors sm:cursor-default cursor-pointer ${mobileActionsTxKey === txKey ? "bg-[var(--surface-light)]/30" : ""}`}
-                      onClick={() => {
-                        // On mobile (< sm), toggle action card
-                        if (window.innerWidth < 640) {
-                          setMobileActionsTxKey(mobileActionsTxKey === txKey ? null : txKey);
-                        }
-                      }}
-                    >
-                      <span className="text-lg flex-shrink-0 w-7 text-center">{icon || "?"}</span>
-                      <div className="flex-1 min-w-0">
-                        <div className="flex items-baseline gap-1.5">
-                          <span className="text-sm truncate">
-                            {tx.description.length > 40
-                              ? tx.description.slice(0, 40) + "..."
-                              : tx.description}
-                          </span>
-                          <span className="text-[11px] text-[var(--text-muted)] flex-shrink-0">
-                            {tx.category}{tx.tag ? ` / ${tx.tag}` : ""}
-                          </span>
-                        </div>
-                        {matchedRule && (
-                          <span
-                            className="inline-flex items-center gap-0.5 mt-0.5 px-1.5 py-px rounded-full bg-violet-500/15 text-violet-400 text-[10px]"
-                            title={`${t("tooltips.matchedRule")}: ${matchedRule.name}`}
-                          >
-                            <Wand2 size={9} />
-                            {matchedRule.name}
-                          </span>
-                        )}
-                      </div>
-                      {/* Action buttons — hidden on small mobile, visible on sm+ */}
-                      <div className="hidden sm:grid grid-cols-3 flex-shrink-0 w-[96px]">
-                        <button
-                          className={`w-[32px] h-[32px] flex items-center justify-center rounded-md transition-colors ${isEditing ? "bg-[var(--primary)]/20 text-[var(--primary)]" : "text-[var(--text-muted)]/40 hover:text-white hover:bg-[var(--surface-light)]"}`}
-                          title={t("tooltips.editCategoryTag")}
-                          onClick={() => setEditingTxKey(isEditing ? null : txKey)}
-                        >
-                          <Tag size={13} />
-                        </button>
-                        <button
-                          className="w-[32px] h-[32px] flex items-center justify-center rounded-md text-[var(--text-muted)]/40 hover:text-white hover:bg-[var(--surface-light)] transition-colors"
-                          title={t("tooltips.splitTransaction")}
-                          onClick={() => setSplittingTransaction(tx)}
-                        >
-                          <Split size={13} />
-                        </button>
-                        {tx.amount < 0 ? (
-                          tx.pending_refund_id ? (
-                            <span className="w-[32px] h-[32px] flex items-center justify-center text-amber-400" title={t("tooltips.pendingRefund")}>
-                              <RefreshCw size={13} className="animate-pulse" />
-                            </span>
-                          ) : (
-                            <button
-                              className="w-[32px] h-[32px] flex items-center justify-center rounded-md text-amber-400/40 hover:text-amber-400 hover:bg-amber-500/20 transition-colors"
-                              title={t("tooltips.markPendingRefund")}
-                              onClick={() => markPendingMutation.mutate(tx)}
-                              disabled={markPendingMutation.isPending}
-                            >
-                              <RefreshCw size={13} />
-                            </button>
-                          )
-                        ) : (
-                          <button
-                            className="w-[32px] h-[32px] flex items-center justify-center rounded-md text-emerald-400/40 hover:text-emerald-400 hover:bg-emerald-500/20 transition-colors"
-                            title={t("tooltips.linkAsRefund")}
-                            onClick={() => setLinkingTransaction(tx)}
-                          >
-                            <Link2 size={13} />
-                          </button>
-                        )}
-                      </div>
-                      <span
-                        className={`text-sm font-semibold flex-shrink-0 tabular-nums text-end w-[80px] ${
-                          isPositive ? "text-emerald-400" : "text-rose-400"
-                        }`}
-                      >
-                        {isPositive ? "+" : ""}
-                        {formatCurrency(tx.amount)}
-                      </span>
-                    </div>
-
-                    {/* Mobile action buttons — shown on tap */}
-                    {mobileActionsTxKey === txKey && (
-                      <div className="sm:hidden flex items-center gap-1.5 mx-2 mb-1 ms-9 p-1.5 rounded-lg bg-[var(--surface-light)]/40 border border-[var(--surface-light)] animate-in fade-in slide-in-from-top-1 duration-150">
-                        <button
-                          className={`flex items-center gap-1.5 px-2.5 py-1.5 rounded-md text-xs font-medium transition-colors ${isEditing ? "bg-[var(--primary)]/20 text-[var(--primary)]" : "text-[var(--text-muted)] hover:text-white hover:bg-[var(--surface-light)]"}`}
-                          onClick={(e) => { e.stopPropagation(); setEditingTxKey(isEditing ? null : txKey); }}
-                        >
-                          <Tag size={13} className="shrink-0" />
-                          {t("common.tag")}
-                        </button>
-                        <button
-                          className="flex items-center gap-1.5 px-2.5 py-1.5 rounded-md text-xs font-medium text-[var(--text-muted)] hover:text-white hover:bg-[var(--surface-light)] transition-colors"
-                          onClick={(e) => { e.stopPropagation(); setSplittingTransaction(tx); }}
-                        >
-                          <Split size={13} className="shrink-0" />
-                          {t("common.split")}
-                        </button>
-                        {tx.amount < 0 ? (
-                          tx.pending_refund_id ? (
-                            <span className="flex items-center gap-1.5 px-2.5 py-1.5 text-xs font-medium text-amber-400">
-                              <RefreshCw size={13} className="shrink-0 animate-pulse" />
-                              {t("common.pending")}
-                            </span>
-                          ) : (
-                            <button
-                              className="flex items-center gap-1.5 px-2.5 py-1.5 rounded-md text-xs font-medium text-amber-400/70 hover:text-amber-400 hover:bg-amber-500/20 transition-colors"
-                              onClick={(e) => { e.stopPropagation(); markPendingMutation.mutate(tx); }}
-                              disabled={markPendingMutation.isPending}
-                            >
-                              <RefreshCw size={13} className="shrink-0" />
-                              {t("common.refund")}
-                            </button>
-                          )
-                        ) : (
-                          <button
-                            className="flex items-center gap-1.5 px-2.5 py-1.5 rounded-md text-xs font-medium text-emerald-400/70 hover:text-emerald-400 hover:bg-emerald-500/20 transition-colors"
-                            onClick={(e) => { e.stopPropagation(); setLinkingTransaction(tx); }}
-                          >
-                            <Link2 size={13} className="shrink-0" />
-                            {t("common.link")}
-                          </button>
-                        )}
-                      </div>
-                    )}
-
-                    {/* Inline tag editing panel */}
-                    {isEditing && categories && (
-                      <div className="mx-2 mb-2 ms-11 rounded-lg border border-[var(--surface-light)] bg-[var(--surface-light)]/20 overflow-hidden">
-                        <div className="flex items-center gap-3 px-3 py-2">
-                          <div className="flex-1 min-w-0">
-                            <label className="text-[10px] uppercase tracking-wider text-[var(--text-muted)] mb-1 block">{t("common.category")}</label>
-                            <SelectDropdown
-                              options={Object.keys(categories).map((c) => ({ label: c, value: c }))}
-                              value={tx.category || ""}
-                              onChange={(cat) => tagMutation.mutate({ tx, category: cat, tag: "" })}
-                              placeholder="Select..."
-                              size="sm"
-                              onCreateNew={async (name) => { await createCategory(name); }}
-                            />
-                          </div>
-                          <div className="flex-1 min-w-0">
-                            <label className="text-[10px] uppercase tracking-wider text-[var(--text-muted)] mb-1 block">{t("common.tag")}</label>
-                            <SelectDropdown
-                              options={
-                                tx.category && categories[tx.category]
-                                  ? categories[tx.category].map((t: string) => ({ label: t, value: t }))
-                                  : []
-                              }
-                              value={tx.tag || ""}
-                              onChange={(tag) => tagMutation.mutate({ tx, category: tx.category, tag })}
-                              placeholder="Select..."
-                              size="sm"
-                              onCreateNew={
-                                tx.category
-                                  ? async (name) => { await createTag(tx.category, name); }
-                                  : undefined
-                              }
-                            />
-                          </div>
-                          <button
-                            className="self-end mb-0.5 px-2.5 py-1 text-[11px] font-medium rounded-md bg-[var(--surface-light)] text-[var(--text-muted)] hover:text-white hover:bg-[var(--surface-light)]/80 transition-colors"
-                            onClick={() => setEditingTxKey(null)}
-                          >
-                            {t("dashboard.done")}
-                          </button>
-                        </div>
-                      </div>
-                    )}
-                  </div>
-                );
-              })}
-            </div>
-          </div>
-        ))}
-
-        {/* Sentinel element for infinite scroll */}
-        <div ref={sentinelRef} className="h-1" />
-
-        {hasMore && (
-          <div className="flex justify-center py-2">
-            <div className="flex items-center gap-2 text-xs text-[var(--text-muted)]">
-              <div className="w-4 h-4 border-2 border-[var(--primary)]/30 border-t-[var(--primary)] rounded-full animate-spin" />
-              {t("common.loading")}
-            </div>
-          </div>
-        )}
-      </div>
-
-      {/* Split Transaction Modal */}
-      {splittingTransaction && (
-        <SplitTransactionModal
-          transaction={splittingTransaction}
-          onClose={() => setSplittingTransaction(null)}
-          onSuccess={() => {
-            setSplittingTransaction(null);
-            queryClient.invalidateQueries({ queryKey: ["transactions"] });
-            invalidateAnalytics();
-          }}
-        />
-      )}
-
-      {/* Link Refund Modal (for positive/income transactions) */}
-      {linkingTransaction && (
-        <LinkRefundModal
-          isOpen={!!linkingTransaction}
-          onClose={() => setLinkingTransaction(null)}
-          refundTransaction={{
-            id: linkingTransaction.unique_id || linkingTransaction.id || 0,
-            source: linkingTransaction.source || "unknown",
-            amount: linkingTransaction.amount,
-            description: linkingTransaction.description,
-          }}
-        />
-      )}
-    </div>
-  );
-}
-
-/* ------------------------------------------------------------------ */
-/*  Section 4 — Tabbed Insights                                       */
-/* ------------------------------------------------------------------ */
-
 /* ================================================================== */
 /*  Main Dashboard Component                                          */
 /* ================================================================== */
+
+// BudgetSpendingGauge → components/dashboard/BudgetSection.tsx
+// RecentTransactionsFeed → components/dashboard/RecentTransactionsSection.tsx
 
 export function Dashboard() {
   const { t } = useTranslation();
